removing lang_options and moving it into mux
