--- plugins/shrs_lang_options/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "shrs_lang_options"
-version = "0.0.1"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-authors = ["nithinmuthukumar"]
-description = "nushell command language for shrs"
-homepage = "https://mrpicklepinosaur.github.io/shrs/"
-categories = ["command-line-interface", "command-line-utilities", "development-tools"]
-keywords = ["shrs", "shell", "posix", "unix"]
-readme = "README.md"
-repository = "https://github.com/MrPicklePinosaur/shrs"
-
-[dependencies]
-shrs = { path = "../../crates/shrs", version = "0.0.1" }
-shrs_mux = {path = "../shrs_mux/", version = "0.0.1"}
-
