--- plugins/shrs_lang_options/examples/basic.rs
@@ -1,3 +1,0 @@
-fn main() {
-    unimplemented!();
-}
