+++ plugins/shrs_mux/examples/basic.rs
@@ -3,7 +3,7 @@ use shrs_mux::MuxPlugin;
 
 fn main() {
     let myshell = ShellBuilder::default()
-        .with_plugin(MuxPlugin)
+        .with_plugin(MuxPlugin::new())
         .build()
         .unwrap();
